requirements.txt
amqp==5.1.1
ansicon==1.89.0
arrow==1.2.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.4
django-compat==1.0.15
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.13.1
jinxed==1.2.0
kombu==5.2.4
prompt-toolkit==3.0.29
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
